requirements/base.txt
Django==4.2
django-jazzmin==3.0.1
djangorestframework==3.15.2
python-decouple==3.8
aiogram==2.14
requests==2.32.3